Correct deleting sub-files

+++ public/app/tpl/us-files.php
@@ -331,7 +331,14 @@
 										echo '<div class="col-4 col-md-2" style="text-align:center;" style="margin-bottom:2px;">';
 										echo '<a alt="Download file" href="?dl=' . $shareId . '"><span style="color:#A2D0C0; margin-right:18px;"><i class="fas fa-angle-double-down"></i></span></a> ';
 										echo '<a alt="Copy link" class="copyText" id="copyTxt" onClick="showAlert()" data-clipboard-text="' . $webPath . '?dl=' . $shareId . '" href="javascript:;"><span style="color:#A4B6DD; margin-right:18px;"><i class="fas fa-link"></i></span></a> ';
-										echo '<a alt="Delete file" href="?page=us-files&del=' . $encFile . '&type=file" class="delete" data-confirm="Are you sure you would like to delete this file?"><span style="color:#D09292;"><i class="far fa-trash-alt"></i></span></a>';
+										
+										// When deleting a file, ensure we are redirected back
+										if(!empty($_GET['browse'])) {
+											echo '<a alt="Delete file" href="?page=us-files&browse=' . $_GET['browse'] . '&del=' . $encFile . '&type=file" class="delete" data-confirm="Are you sure you would like to delete this file?"><span style="color:#D09292;"><i class="far fa-trash-alt"></i></span></a>';
+										}else{
+											echo '<a alt="Delete file" href="?page=us-files&del=' . $encFile . '&type=file" class="delete" data-confirm="Are you sure you would like to delete this file?"><span style="color:#D09292;"><i class="far fa-trash-alt"></i></span></a>';
+										}
+										
 										echo '</div>';
 									}
 								}
